+++ usage/help.go
@@ -55,10 +55,6 @@ $ step help ssh
 				Name:  "report",
 				Usage: "Writes a JSON report to the HTML docs directory.",
 			},
-			cli.BoolFlag{
-				Name:  "hugo",
-				Usage: "Writes hugo (vs jekyll) compatible markdown files",
-			},
 		},
 	}
 }